Runescape/original_failure/main.py
#!/usr/bin/env python3
"""
Runescape Bot - Main Application
Isolated environment for bot development
"""

import os
import sys
import logging
from flask import Flask, jsonify, request
from dotenv import load_dotenv

# Load environment variables
load_dotenv()

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler('/app/logs/bot.log'),
        logging.StreamHandler(sys.stdout)
    ]
)

logger = logging.getLogger(__name__)

# Initialize Flask app
app = Flask(__name__)

@app.route('/health')
def health_check():
    """Health check endpoint"""
    return jsonify({
        'status': 'healthy',
        'service': 'runescape-bot',
        'version': os.getenv('BOT_VERSION', '1.0.0')
    })

@app.route('/status')
def bot_status():
    """Get bot status"""
    return jsonify({
        'bot_name': os.getenv('BOT_NAME', 'RunescapeBot'),
        'environment': os.getenv('BOT_ENV', 'development'),
        'database_url': os.getenv('DATABASE_URL', 'not_configured'),
        'redis_url': os.getenv('REDIS_URL', 'not_configured')
    })

@app.route('/start', methods=['POST'])
def start_bot():
    """Start the bot"""
    try:
        # Add your bot logic here
        logger.info("Bot start requested")
        return jsonify({'message': 'Bot started successfully'})
    except Exception as e:
        logger.error(f"Error starting bot: {e}")
        return jsonify({'error': str(e)}), 500

@app.route('/stop', methods=['POST'])
def stop_bot():
    """Stop the bot"""
    try:
        logger.info("Bot stop requested")
        return jsonify({'message': 'Bot stopped successfully'})
    except Exception as e:
        logger.error(f"Error stopping bot: {e}")
        return jsonify({'error': str(e)}), 500

if __name__ == '__main__':
    logger.info("Starting Runescape Bot...")
    logger.info(f"Environment: {os.getenv('BOT_ENV', 'development')}")
    logger.info(f"Database: {os.getenv('DATABASE_URL', 'not_configured')}")
    
    app.run(
        host=os.getenv('API_HOST', '0.0.0.0'),
        port=int(os.getenv('API_PORT', 8080)),
        debug=os.getenv('BOT_ENV') == 'development'
    )
