Runescape/original_failure/requirements.txt
# Core dependencies
requests==2.31.0
selenium==4.15.2
beautifulsoup4==4.12.2
pillow==10.1.0
numpy==1.24.3
opencv-python==4.8.1.78

# Database
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
alembic==1.12.1

# Redis
redis==5.0.1

# Web framework
flask==3.0.0
flask-cors==4.0.0

# Configuration
python-dotenv==1.0.0
pydantic==2.5.0

# Logging
loguru==0.7.2

# Image processing
pyautogui==0.9.54
pytesseract==0.3.10

# Async support
asyncio==3.4.3
aiohttp==3.9.1

# Development
pytest==7.4.3
black==23.11.0
flake8==6.1.0
